JAMES-2813 Split TaskManager API from Memory implementation
