+++ server/task-api/src/main/java/org/apache/james/task/TaskManager.java
