+++ server/task-api/src/test/java/org/apache/james/task/MemoryReferenceTask.java
