+++ server/task/src/test/java/org/apache/james/task/MemoryTaskManagerTest.java
@@ -19,8 +19,6 @@
 
 package org.apache.james.task;
 
-import org.apache.james.task.eventsourcing.Hostname;
-
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.extension.ExtendWith;
